+++ src/renderer/MirrorMapRenderer.cpp
@@ -115,8 +115,8 @@ void MirrorMapRenderer::updateView(const RenderContext& ctx)
 
     const auto& res = ctx.m_resolution;
 
-    int w = ctx.m_assets.mirrorReflectionBufferScale * res.x;
-    int h = ctx.m_assets.mirrorReflectionBufferScale * res.y;
+    int w = (int)(ctx.m_assets.mirrorReflectionBufferScale * res.x);
+    int h = (int)(ctx.m_assets.mirrorReflectionBufferScale * res.y);
 
     if (m_squareAspectRatio) {
         h = w;
